requirements.txt
asn1crypto==0.24.0
certifi==2018.1.18
cffi==1.11.4
chardet==3.0.4
click==6.7
cryptography==2.1.4
enum34==1.1.6
Flask==0.12.2
gunicorn==19.7.1
httplib2==0.10.3
idna==2.6
ipaddress==1.0.19
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
oauth2==1.9.0.post1
oauthlib==2.0.6
pycparser==2.18
PyJWT==1.5.3
pyOpenSSL==17.5.0
python-flickr==0.3.2
pytz==2017.3
requests==2.18.4
requests-oauthlib==0.8.0
requests-toolbelt==0.8.0
simplejson==3.13.2
six==1.11.0
twilio==6.10.1
urllib3==1.22
Werkzeug==0.14.1
